+++ src/Head.php
@@ -4,25 +4,62 @@ namespace ProtoneMedia\Splade;
 
 use Illuminate\Contracts\Support\Arrayable;
 use Illuminate\Contracts\Support\Htmlable;
+use Illuminate\Support\Collection;
 use Illuminate\Support\HtmlString;
-use Illuminate\Support\Str;
+use Illuminate\Support\Traits\Macroable;
 use JsonSerializable;
+use ProtoneMedia\Splade\SEO\OpenGraph;
+use ProtoneMedia\Splade\SEO\Twitter;
 
 class Head implements Arrayable, JsonSerializable
 {
-    private string $title;
+    use Macroable;
+    use OpenGraph;
+    use Twitter;
 
-    private array $meta = [];
+    private bool $autoFillOverwrite = false;
+
+    private string $title = '';
+
+    private ?string $canonical = null;
+
+    private Collection $meta;
 
     /**
      * Creates a new instance and sets the defaults from the config.
      */
     public function __construct()
     {
+        $this->meta = new Collection;
+
+        $this->fillOpenGraphDefaults();
+        $this->fillTwitterDefaults();
+
         $this
-            ->title(config('splade.seo.defaults.title'))
-            ->description(config('splade.seo.defaults.description'))
-            ->keywords(config('splade.seo.defaults.keywords'));
+            ->title(config('splade.seo.defaults.title') ?: '')
+            ->description(config('splade.seo.defaults.description') ?: '')
+            ->keywords(config('splade.seo.defaults.keywords') ?: '');
+
+        if (config('splade.seo.auto_canonical_link')) {
+            $this->canonical(request()->fullUrl());
+        }
+
+        $this->autoFillOverwrite = true;
+    }
+
+    /**
+     * Use the traits the fill the meta properties with the default data.
+     *
+     * @return self
+     */
+    private function autoFill(): self
+    {
+        if ($this->autoFillOverwrite) {
+            $this->autoFillOpenGraph();
+            $this->autoFillTwitter();
+        }
+
+        return $this;
     }
 
     /**
@@ -34,13 +71,34 @@ class Head implements Arrayable, JsonSerializable
      */
     public function title(string $title, bool $withPrefixAndSuffix = true): self
     {
-        $this->title = trim(implode(' ', [
-            $withPrefixAndSuffix ? trim(config('splade.seo.title_prefix')) : '',
+        $title     = trim($title);
+        $prefix    = trim(config('splade.seo.title_prefix'));
+        $suffix    = trim(config('splade.seo.title_suffix'));
+        $separator = trim(config('splade.seo.title_separator'));
 
-            trim($title),
+        $withPrefix = $withPrefixAndSuffix && $prefix && $prefix !== $title;
+        $withSuffix = $withPrefixAndSuffix && $suffix && $suffix !== $title;
 
-            $withPrefixAndSuffix ? trim(config('splade.seo.title_suffix')) : '',
+        $this->title = implode(' ', array_filter([
+            $withPrefix ? $prefix : false,
+            $withPrefix ? $separator : false,
+            $title,
+            $withSuffix ? $separator : false,
+            $withSuffix ? $suffix : false,
         ]));
+
+        return $this->autoFill();
+    }
+
+    /**
+     * Setter for the canonical URL.
+     *
+     * @param  string  $url
+     * @return $this
+     */
+    public function canonical(string $url): self
+    {
+        $this->canonical = $url;
 
         return $this;
     }
@@ -53,7 +111,9 @@ class Head implements Arrayable, JsonSerializable
      */
     public function description(string $description): self
     {
-        return $this->metaByName('description', trim($description));
+        $this->metaByName('description', trim($description));
+
+        return $this->autoFill();
     }
 
     /**
@@ -73,7 +133,9 @@ class Head implements Arrayable, JsonSerializable
                 ->implode(', ');
         }
 
-        return $this->metaByName('keywords', $keywords);
+        $this->metaByName('keywords', $keywords);
+
+        return $this->autoFill();
     }
 
     /**
@@ -81,10 +143,21 @@ class Head implements Arrayable, JsonSerializable
      *
      * @param  string  $name
      * @param  string  $content
+     * @param  bool  $replace
      * @return $this
      */
-    public function metaByName(string $name, string $content): self
+    public function metaByName(string $name, string $content, bool $replace = true): self
     {
+        if ($replace) {
+            $this->removeMeta(['name' => $name]);
+        }
+
+        $content = trim($content);
+
+        if (!$content) {
+            return $this;
+        }
+
         return $this->meta(['name' => $name, 'content' => $content]);
     }
 
@@ -93,32 +166,86 @@ class Head implements Arrayable, JsonSerializable
      *
      * @param  string  $property
      * @param  string  $content
+     * @param  bool  $replace
      * @return $this
      */
-    public function metaByProperty(string $property, string $content): self
+    public function metaByProperty(string $property, string $content, bool $replace = true): self
     {
+        if ($replace) {
+            $this->removeMeta(['property' => $property]);
+        }
+
+        $content = trim($content);
+
+        if (!$content) {
+            return $this;
+        }
+
         return $this->meta(['property' => $property, 'content' => $content]);
     }
 
     /**
-     * Sets a meta tag by the given attributes.
+     * Adds a meta tag by the given attributes.
      *
      * @param  array  $attributes
      * @return $this
      */
     public function meta(array $attributes): self
     {
-        if ($value = data_get($attributes, 'name')) {
-            $key = "name.{$value}";
-        } elseif ($value = data_get($attributes, 'property')) {
-            $key = "property.{$value}";
-        } else {
-            $key = Str::uuid()->toString();
-        }
-
-        $this->meta[$key] = new Meta($key, $attributes);
+        $this->meta->push(new Meta($attributes));
 
         return $this;
+    }
+
+    /**
+     * Remove a meta tag that matches the given attributes
+     *
+     * @param  array  $attributes
+     * @return void
+     */
+    public function removeMeta(array $attributes)
+    {
+        $this->meta = $this->meta->reject(function (Meta $meta) use ($attributes) {
+            return $meta->hasAllAttributes($attributes);
+        });
+
+        return $this;
+    }
+
+    /**
+     * Get a Meta instance by name.
+     *
+     * @param  string  $name
+     * @return \Illuminate\Support\Collection
+     */
+    public function getMetaByName(string $name): Collection
+    {
+        return $this->meta->filter(function (Meta $meta) use ($name) {
+            return $meta->name === $name;
+        });
+    }
+
+    /**
+     * Get a Meta instance by property.
+     *
+     * @param  string  $property
+     * @return \Illuminate\Support\Collection
+     */
+    public function getMetaByProperty(string $property): Collection
+    {
+        return $this->meta->filter(function (Meta $meta) use ($property) {
+            return $meta->property === $property;
+        });
+    }
+
+    /**
+     * Getter for the title.
+     *
+     * @return string
+     */
+    public function getTitle(): string
+    {
+        return $this->title;
     }
 
     /**
@@ -138,9 +265,14 @@ class Head implements Arrayable, JsonSerializable
      */
     private function renderMeta(): string
     {
-        return collect($this->meta)->map(function (Meta $meta) {
-            return $meta->render();
-        })->implode(PHP_EOL);
+        return $this->meta
+            ->map(fn (Meta $meta) => $meta->render())
+            ->when($this->canonical, function (Collection $collection, string $href) {
+                $collection->prepend(
+                    "<link rel=\"canonical\" href=\"{$href}\">"
+                );
+            })
+            ->implode(PHP_EOL);
     }
 
     /**
@@ -161,8 +293,9 @@ class Head implements Arrayable, JsonSerializable
     public function toArray(): array
     {
         return [
-            'title' => $this->title,
-            'meta'  => array_values($this->meta),
+            'canonical' => $this->canonical,
+            'meta'      => $this->meta->values()->all(),
+            'title'     => $this->title,
         ];
     }