+++ src/Http/SpladeMiddleware.php
@@ -7,13 +7,16 @@ use Illuminate\Contracts\Session\Session;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response as LaravelResponse;
+use Illuminate\Routing\UrlGenerator;
 use Illuminate\Support\Facades\Blade;
 use Illuminate\Support\Str;
 use Illuminate\Validation\ValidationException;
 use Illuminate\View\View;
+use ProtoneMedia\Splade\Facades\Splade;
 use ProtoneMedia\Splade\SpladeCore;
 use ProtoneMedia\Splade\Ssr;
 use Symfony\Component\HttpFoundation\BinaryFileResponse;
+use Symfony\Component\HttpFoundation\RedirectResponse;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\HttpFoundation\StreamedResponse;
 
@@ -56,6 +59,11 @@ class SpladeMiddleware
 
         // Gather the required meta data for the app.
         $spladeData = $this->spladeData($request->session());
+
+        // The response should redirect away from the Splade app.
+        if ($redirect = $this->shouldRedirectsAway($response)) {
+            return $redirect;
+        }
 
         // If the response is a redirect, put the toasts into the session
         // so they won't be lost in the next request.
@@ -269,10 +277,74 @@ class SpladeMiddleware
                 $session->pull(static::FLASH_TOASTS, []),
                 $this->splade->getToasts(),
             ),
-
             'preventRefresh' => $this->splade->dontRefreshPage(),
             'lazy'           => $this->splade->isLazyRequest(),
         ];
+    }
+
+    /**
+     * Returns a boolean whether the Target URL is outside of the app.
+     *
+     * @param  \Symfony\Component\HttpFoundation\Response  $redirect
+     * @return bool
+     */
+    private function shouldRedirectsAway(Response $response): bool|Response
+    {
+        if ($response->headers->has(SpladeCore::HEADER_REDIRECT_AWAY)) {
+            // This is already a redirect to an outside target.
+            return $response;
+        }
+
+        if (!$response instanceof RedirectResponse) {
+            // No redirect at all.
+            return false;
+        }
+
+        $targetUrl = $response->getTargetUrl();
+
+        if (Str::startsWith($targetUrl, '/')) {
+            // Relative URL.
+            return false;
+        }
+
+        /** @var UrlGenerator $urlGenerator */
+        $urlGenerator = app('url');
+
+        $appUrl = $urlGenerator->format(
+            $urlGenerator->formatRoot($urlGenerator->formatScheme()) ?: config('app.url'),
+            '/'
+        );
+
+        if ($this->urlToHostAndPort($appUrl) === $this->urlToHostAndPort($targetUrl)) {
+            // A redirect inside the app.
+            return false;
+        }
+
+        return Splade::redirectAway($targetUrl);
+    }
+
+    /**
+     * Maps a full URL to a host:port formatted string.
+     *
+     * @param  string  $url
+     * @return string
+     */
+    public static function urlToHostAndPort(string $url): string
+    {
+        if (!parse_url($url, PHP_URL_HOST)) {
+            $url = "http://{$url}";
+        }
+
+        $parsed = parse_url($url);
+
+        $host = $parsed['host'] ?? 'host';
+        $port = $parsed['port'] ?? null;
+
+        if (!$port) {
+            $port = $parsed['scheme'] === 'http' ? 80 : 443;
+        }
+
+        return "{$host}:{$port}";
     }
 
     /**