+++ src/SpladeCore.php
@@ -4,10 +4,12 @@ namespace ProtoneMedia\Splade;
 
 use Closure;
 use Illuminate\Foundation\Exceptions\Handler;
+use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Support\Arr;
 use Illuminate\Validation\ValidationException;
 use ProtoneMedia\Splade\Http\ResolvableData;
+use Symfony\Component\HttpFoundation\Response;
 use Throwable;
 
 class SpladeCore
@@ -19,6 +21,8 @@ class SpladeCore
     const HEADER_PREVENT_REFRESH = 'X-Splade-Prevent-Refresh';
 
     const HEADER_LAZY = 'X-Splade-Lazy';
+
+    const HEADER_REDIRECT_AWAY = 'X-Splade-Redirect-Away';
 
     const MODAL_TYPE_MODAL = 'modal';
 
@@ -361,5 +365,23 @@ class SpladeCore
             'slideover' => static::MODAL_TYPE_SLIDEOVER,
             default     => static::MODAL_TYPE_MODAL
         };
+    }
+
+    /**
+     * Returns a JSON response that indicates that the Splade frontend
+     * should redirect to an external URL.
+     *
+     * @param  string  $targetUrl
+     * @return \Symfony\Component\HttpFoundation\Response
+     */
+    public function redirectAway(string $targetUrl): Response
+    {
+        if (!$this->isSpladeRequest()) {
+            return redirect()->away($targetUrl);
+        }
+
+        return new JsonResponse(null, 409, [
+            static::HEADER_REDIRECT_AWAY => $targetUrl,
+        ]);
     }
 }